rag/mistral/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.3.0

# LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-mistralai>=0.1.0
langchain-huggingface>=0.0.3

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Document processing
pypdf>=3.17.0
unstructured>=0.11.0
unstructured[md]>=0.11.0
docling>=1.0.0

# Embeddings
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4

# API
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
aiofiles>=23.2.1

# Caching
redis>=5.0.1

# Utils
numpy>=1.24.3
pandas>=2.1.4
tiktoken>=0.5.2

# Development
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.12.0
flake8==7.0.0
mypy==1.7.1